#228 Refactored missed namespaces
